+++ packages/docs/.gitignore
