+++ packages/docs/components/Examples.tsx
@@ -1,11 +1,10 @@
-import * as React from "react";
-import { Example } from "../components/Example";
 import { Basic } from "../examples/Basic";
 import { Browser } from "../examples/Browser";
 import { Draggable } from "../examples/Draggable";
 import { Dropdown } from "../examples/Dropdown";
 import { Nested } from "../examples/Nested";
 import { Overflow } from "../examples/Overflow";
+import { Example } from "./Example";
 
 export function Examples() {
   return (
