+++ packages/docs/examples/Dropdown.tsx
@@ -1,17 +1,14 @@
-import * as React from "react";
 import { TabPanel, useTabs } from "react-headless-tabs";
 import { TabSelector } from "./TabSelector";
 
-export function Basic() {
-  const [selectedTab, setSelectedTab] = useTabs([
-    "account",
-    "company",
-    "team",
-    "billing",
-  ]);
+export function Dropdown() {
+  const [selectedTab, setSelectedTab] = useTabs(
+    ["account", "company", "team", "billing"],
+    null
+  );
 
   return (
-    <>
+    <div className="relative">
       <nav className="flex border-b border-gray-300">
         <TabSelector
           isActive={selectedTab === "account"}
@@ -38,12 +35,24 @@ export function Basic() {
           Billing
         </TabSelector>
       </nav>
-      <div className="p-4">
+      <div
+        className="absolute left-0 ml-4 p-4 w-full border border-t-0 border-gray-300 max-w-lg bg-gray-100 bg-opacity-90 transition-transform origin-top duration-100 ease-in-out"
+        style={{ transform: selectedTab ? "scaleY(1)" : "scaleY(0)" }}
+      >
         <TabPanel hidden={selectedTab !== "account"}>My Account</TabPanel>
         <TabPanel hidden={selectedTab !== "company"}>Company</TabPanel>
         <TabPanel hidden={selectedTab !== "team"}>Team Members</TabPanel>
         <TabPanel hidden={selectedTab !== "billing"}>Billing</TabPanel>
+        <button
+          className="absolute top-0 right-0 mt-2 mr-2 text-gray-600"
+          onClick={() => setSelectedTab(null)}
+        >
+          Close
+        </button>
       </div>
-    </>
+      <div className="overflow-hidden p-4 prose prose-lg text-gray-600 mx-auto">
+        <p>Incredibly amazing awesome content.</p>
+      </div>
+    </div>
   );
 }
