+++ packages/docs/next.config.js
