+++ packages/docs/package.json
