+++ packages/docs/tailwind.config.js
