+++ packages/docs/tsconfig.json
