+++ packages/docs/vite.config.ts
