+++ packages/lib/package.json
