+++ packages/lib/tsconfig.json
