+++ packages/lib/vite.config.ts
