+++ pnpm-workspace.yaml
@@ -1,4 +1,3 @@
 packages:
-  - "docs/**"
-  - "lib/**"
+  - "packages/**"
   - "!**/test/**"
